requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
cryptography==2.7
idna==2.8
numpy==1.18.3
pandas==1.0.3
psycopg2-binary==2.8.4
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.24.2
win-inet-pton==1.1.0
wincertstore==0.2
